[Python] Compiling python files to fix tests
This commit allows the "python_parser" test to pass, even if pinocchio
is not installed, by compiling all python scripts into
CMAKE_BUILD_DIR/bindings/python/pinocchio, where the .so already is, and
add this directory to the PYTHONPATH for this test.

Compiling python files also allows to detect errors in them, instead of
blindly installing them on PYTHONPATH

We could discuss about installing those compiled versions, as it could
speed up the library use, as normal users usually can't write on the
system's PYTHONPATH, and therefore the interpreter can't
compile-and-install the source files on the fly, so it has to compile
them each and every time.