Always add /bigobj for compilation units that link pinocchio
Pinocchio with its heavely templated code create a huge number of
sections, and to avoid C1128 errors, the simpler solution is to
always pass /bigobj on MSVC to compilation units that link pinocchio.
The only downside of this approach is that library generates with this
option cannot be linked with linkers shipped with Visual Studio 2005 or
earlier, that I think it is not a problem in practice.