Turn off automatic documentation generation
The documentation is generated automatically if Doxygen is installed,
which takes a considerable amount of time. Most users are likely to use
the online-hosted Doxygen by default and so setting the default for
INSTALL_DOCUMENTATION to false should be fine (as is done in EigenPy).
For distribution/binaries, we save ~36-77 MB of installed files.