SO3 diff_impl: use quat instead of matrix
For implementation of difference between two SO3 values, it is more
accurate to use quaternion multiplication instead of the corresponding
rotation matrix form.

However, with the new diff_impl, integrate(q0, difference(q0,q1)) is less
similar to the expected result q1. This is probably because the current
implementation of integrate also suffers from floating point error. The
threshold for the integrate_difference_test needs to be changed slightly
to accommodate the new implementation of difference.