cmake: Always suppose hpp-fcl Python binding are present when hpp-fcl C++ library is found
Pinocchio Python binding needs hpp-fcl Python binding for
collision/contact related algorithms.

We allow with the `BUILD_WITH_HPP_FCL_PYTHON_BINDINGS` CMake variable to build
Pinocchio C++ library with hpp-fcl but to ignore hpp-fcl in the Python
bindings if the hpp-fcl Python bindings are not installed.

This create the following issues:
- To compute `BUILD_WITH_HPP_FCL_PYTHON_BINDINGS` we must call the
  Python interpreter. This create some issue when cross compiling.
- We never test this case in our CI, so we don't know if this option is
  working.

This PR remove this option to simplify the build.
In the case an user build Pinocchio Python binding with hpp-fcl
C++ library but without hpp-fcl binding, the Python interpreter will
raise an ImportError.